% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{accession<-,ProjectSummary-method}
\alias{accession<-,ProjectSummary-method}
\title{Replaces a project accession}
\usage{
\S4method{accession}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the accession}
}
\description{
Replaces a project accession
}
\author{
Jose A. Dianes
}

