% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-ContactDetail.R
\docType{methods}
\name{affiliation,ContactDetail-method}
\alias{affiliation,ContactDetail-method}
\title{Returns the affiliation of a contact}
\usage{
\S4method{affiliation}{ContactDetail}(object)
}
\arguments{
\item{object}{a ContactDetail}
}
\value{
the affiliation
}
\description{
Returns the affiliation of a contact
}

