% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-ContactDetail.R
\docType{methods}
\name{affiliation<-,ContactDetail-method}
\alias{affiliation<-,ContactDetail-method}
\title{Replaces the affiliation of a contact}
\usage{
\S4method{affiliation}{ContactDetail}(object) <- value
}
\arguments{
\item{object}{a ContactDetail}

\item{value}{the new affiliation}
}
\description{
Replaces the affiliation of a contact
}

