% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{as.data.frame,Cluster-method}
\alias{as.data.frame,Cluster-method}
\title{Returns a data frame from a Cluster}
\usage{
\S4method{as.data.frame}{Cluster}(object, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{row.names}{optional row names}

\item{optional}{optional}

\item{x}{The cluster}
}
\value{
The cluster as a data frame
}
\description{
Returns a data frame from a Cluster
}
\details{
TODO
}
\author{
Jose A. Dianes
}

