% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.data.frame,ClusterSearchResults-method}
\alias{as.data.frame,ClusterSearchResults-method}
\title{Returns a data frame from a ClusterSearchResult results}
\usage{
\S4method{as.data.frame}{ClusterSearchResults}(object, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{row.names}{optional row names}

\item{optional}{optional}

\item{x}{The cluster search results}
}
\value{
The cluster search results as a data frame
}
\description{
Returns a data frame from a ClusterSearchResult results
}
\details{
TODO
}
\author{
Jose A. Dianes
}

