% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{as.data.frame,ProjectSummary-method}
\alias{as.data.frame,ProjectSummary-method}
\title{Returns a data frame from ProjectSummary inputs}
\usage{
\S4method{as.data.frame}{ProjectSummary}(object, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{row.names}{optional row names}

\item{x}{The project summaries}
}
\value{
The project summaries as a data frame
}
\description{
Returns a data frame from ProjectSummary inputs
}
\details{
TODO
}
\author{
Jose A. Dianes
}

