% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{as.data.frame,ProjectSummaryList-method}
\alias{as.data.frame,ProjectSummaryList-method}
\title{Returns a data frame from a ProteinDetailList instance}
\usage{
\S4method{as.data.frame}{ProjectSummaryList}(object, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{row.names}{optional row names}

\item{optional}{optional}

\item{x}{The ProjectSummaryList instance}
}
\value{
The page of project summaries as a data frame
}
\description{
Returns a data frame from a ProteinDetailList instance
}
\details{
TODO
}
\author{
Jose A. Dianes
}

