% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.data.frame,ProteinDetailList-method}
\alias{as.data.frame,ProteinDetailList-method}
\title{Returns a data frame from a ProteinDetailList instance}
\usage{
\S4method{as.data.frame}{ProteinDetailList}(object, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{row.names}{optional row names}

\item{optional}{optional}

\item{x}{The ProteinDetailList instance}
}
\value{
The page of protein details as a data frame
}
\description{
Returns a data frame from a ProteinDetailList instance
}
\details{
TODO
}
\author{
Jose A. Dianes
}

