% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions-FileDetail.R
\name{as.data.frame.FileDetail}
\alias{as.data.frame.FileDetail}
\title{Returns a data frame from FileDetail inputs}
\usage{
\method{as.data.frame}{FileDetail}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{The file details}

\item{row.names}{optional row names}
}
\value{
The file details as a data frame
}
\description{
Returns a data frame from FileDetail inputs
}
\details{
TODO
}
\author{
Jose A. Dianes
}

