% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\name{as.data.frame.ProjectSummary}
\alias{as.data.frame.ProjectSummary}
\title{Returns a data frame from ProjectSummary inputs}
\usage{
as.data.frame.ProjectSummary(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{The project summaries}

\item{row.names}{optional row names}
}
\value{
The project summaries as a data frame
}
\description{
Returns a data frame from ProjectSummary inputs
}
\details{
TODO
}
\author{
Jose A. Dianes
}

