% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.ProteinDetail}
\alias{as.data.frame.ProteinDetail}
\title{Returns a data frame from ProteinDetail inputs}
\usage{
\method{as.data.frame}{ProteinDetail}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{The protein detail inputs}

\item{row.names}{optional row names}
}
\value{
The protein identification details as a data frame
}
\description{
Returns a data frame from ProteinDetail inputs
}
\details{
TODO
}
\author{
Jose A. Dianes
}

