% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\name{as.data.frame.PsmDetail}
\alias{as.data.frame.PsmDetail}
\title{Returns a data frame from PsmDetail inputs}
\usage{
\method{as.data.frame}{PsmDetail}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{The PSM detail inputs}

\item{row.names}{optional row names}
}
\value{
The PSM details as a data frame
}
\description{
Returns a data frame from PsmDetail inputs
}
\details{
TODO
}
\author{
Jose A. Dianes
}

