% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{assay.accession,AssayDetail-method}
\alias{assay.accession,AssayDetail-method}
\title{Returns an assay accession}
\usage{
\S4method{assay.accession}{AssayDetail}(object)
}
\arguments{
\item{object}{a AssayDetail}
}
\value{
the accession
}
\description{
Returns an assay accession
}
\author{
Jose A. Dianes
}

