% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{assay.accession,FileDetail-method}
\alias{assay.accession,FileDetail-method}
\title{Returns a file assay accession}
\usage{
\S4method{assay.accession}{FileDetail}(object)
}
\arguments{
\item{object}{a FileDetail}
}
\value{
the assay accession
}
\description{
Returns a file assay accession
}
\author{
Jose A. Dianes
}

