% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{assay.accession,ProteinDetail-method}
\alias{assay.accession,ProteinDetail-method}
\title{Returns a protein assay accession}
\usage{
\S4method{assay.accession}{ProteinDetail}(object)
}
\arguments{
\item{object}{a ProjectDetail}
}
\value{
the assay accession
}
\description{
Returns a protein assay accession
}
\author{
Jose A. Dianes
}

