% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{assay.accession,PsmDetail-method}
\alias{assay.accession,PsmDetail-method}
\title{Returns a PSM assay.accession}
\usage{
\S4method{assay.accession}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the assay.accession
}
\description{
Returns a PSM assay.accession
}
\author{
Jose A. Dianes
}

