% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{assay.accession<-,AssayDetail-method}
\alias{assay.accession<-,AssayDetail-method}
\title{Replaces an assay accession}
\usage{
\S4method{assay.accession}{AssayDetail}(object) <- value
}
\arguments{
\item{object}{a AssayDetail}

\item{value}{the accession}
}
\description{
Replaces an assay accession
}
\author{
Jose A. Dianes
}

