% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{assay.accession<-,FileDetail-method}
\alias{assay.accession<-,FileDetail-method}
\title{Replaces a file assay accession}
\usage{
\S4method{assay.accession}{FileDetail}(object) <- value
}
\arguments{
\item{object}{a FileDetail}

\item{value}{the assay accession}
}
\description{
Replaces a file assay accession
}
\author{
Jose A. Dianes
}

