% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{assay.accession<-,ProteinDetail-method}
\alias{assay.accession<-,ProteinDetail-method}
\title{Replaces a protein assay accession}
\usage{
\S4method{assay.accession}{ProteinDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the assay accession}
}
\description{
Replaces a protein assay accession
}
\author{
Jose A. Dianes
}

