% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{average.precursor.charge,Cluster-method}
\alias{average.precursor.charge,Cluster-method}
\title{Returns a cluster average.precursor.charge}
\usage{
\S4method{average.precursor.charge}{Cluster}(object)
}
\arguments{
\item{object}{a Cluster}
}
\value{
the average.precursor.charge
}
\description{
Returns a cluster average.precursor.charge
}
\author{
Jose A. Dianes
}

