% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{calculated.mz,PsmDetail-method}
\alias{calculated.mz,PsmDetail-method}
\title{Returns a PSM calculated.mz}
\usage{
\S4method{calculated.mz}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the calculated.mz
}
\description{
Returns a PSM calculated.mz
}
\author{
Jose A. Dianes
}

