% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{charge,PsmDetail-method}
\alias{charge,PsmDetail-method}
\title{Returns a PSM charge}
\usage{
\S4method{charge}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the charge
}
\description{
Returns a PSM charge
}
\author{
Jose A. Dianes
}

