% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{charge<-,PsmDetail-method}
\alias{charge<-,PsmDetail-method}
\title{Replaces a PSM charge}
\usage{
\S4method{charge}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the charge}
}
\description{
Replaces a PSM charge
}
\author{
Jose A. Dianes
}

