% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{class}
\name{ClusterSearchResults-class}
\alias{ClusterSearchResults-class}
\title{ClusterSearchResults represents a PRIDE Cluster project summary result set}
\description{
ClusterSearchResults represents a PRIDE Cluster project summary result set
}

