% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ContactDetail-class}
\alias{ContactDetail-class}
\title{ContactDetail represents a PRIDE Archive assay's contact}
\description{
ContactDetail represents a PRIDE Archive assay's contact
}

