% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-ContactDetail.R
\name{ContactDetail}
\alias{ContactDetail}
\title{Constructor for ContactDetail}
\usage{
ContactDetail(title, first.name, last.name, email, affiliation)
}
\arguments{
\item{title}{the title of the contact}

\item{first.name}{the first name of the contact}

\item{last.name}{the last name of the contact}

\item{affiliation}{the affiliation of the contact}
}
\description{
Constructor for ContactDetail
}

