% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions-FileDetail.R
\name{count.assay.FileDetail}
\alias{count.assay.FileDetail}
\title{Returns the number of files associated with an assay}
\usage{
count.assay.FileDetail(assay.project)
}
\arguments{
\item{assay.accession}{the assay accession}
}
\value{
The count of files
}
\description{
Returns the number of files associated with an assay
}
\details{
TODO
}
\author{
Jose A. Dianes
}

