% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assay.R
\name{count.AssayDetail}
\alias{count.AssayDetail}
\title{Returns the number of assays associated with a project}
\usage{
count.AssayDetail(project.accession)
}
\arguments{
\item{project.accession}{the project accession}
}
\value{
The count of assays
}
\description{
Returns the number of assays associated with a project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

