% Generated by roxygen2 (4.0.1): do not edit by hand
\name{count.ProjectSummary}
\alias{count.ProjectSummary}
\title{Returns the number of public projects in PRIDE Archive}
\usage{
count.ProjectSummary()
}
\value{
The count of projects
}
\description{
Returns the number of public projects in PRIDE Archive
}
\details{
TODO
}
\author{
Jose A. Dianes
}

