% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count.ProteinDetail}
\alias{count.ProteinDetail}
\title{Returns the number of proteins for a particual public project}
\usage{
count.ProteinDetail(project.accession)
}
\arguments{
\item{project.accession}{The project accession to count proteins from}
}
\value{
The count of proteins
}
\description{
Returns the number of proteins for a particual public project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

