% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{create_random_name}
\alias{create_random_name}
\title{Creates random tmp filenames}
\usage{
create_random_name(length = 20, suffix = ".csv")
}
\arguments{
\item{length}{A number}

\item{suffix}{A string}
}
\value{
A string; random filename
}
\description{
Creates random tmp filenames
}
\examples{
create_random_name()
create_random_name(length=30, suffix = ".txt")
}
