% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{description,ProjectSummary-method}
\alias{description,ProjectSummary-method}
\title{Returns a project description}
\usage{
\S4method{description}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project description
}
\description{
Returns a project description
}
\author{
Jose A. Dianes
}

