% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{description<-,ProjectSummary-method}
\alias{description<-,ProjectSummary-method}
\title{Replaces a project description}
\usage{
\S4method{description}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the project description}
}
\description{
Replaces a project description
}
\author{
Jose A. Dianes
}

