% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{description<-,ProteinDetail-method}
\alias{description<-,ProteinDetail-method}
\title{Replaces a protein description}
\usage{
\S4method{description}{ProteinDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the description}
}
\description{
Replaces a protein description
}
\author{
Jose A. Dianes
}

