% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{download.link,FileDetail-method}
\alias{download.link,FileDetail-method}
\title{Returns a file download link}
\usage{
\S4method{download.link}{FileDetail}(object)
}
\arguments{
\item{object}{a FileDetail}
}
\value{
the download.link
}
\description{
Returns a file download link
}
\author{
Jose A. Dianes
}

