% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{draw_primers}
\alias{draw_primers}
\title{Prepare a (nearly) optimal primer coverage of the sample set}
\usage{
draw_primers(
  fasta_file,
  primer_length = 20,
  coverage = 0.9,
  minimum_primer_group_size = 10,
  minimum_sequence_group_size = 10,
  draws = 100
)
}
\arguments{
\item{fasta_file}{A string}

\item{primer_length}{A number}

\item{coverage}{A number}

\item{minimum_primer_group_size}{A number}

\item{minimum_sequence_group_size}{A number}

\item{draws}{A number}
}
\value{
A list containing a sequence conversion table and 
        a primer coverage table
}
\description{
Prepare a (nearly) optimal primer coverage of the sample set
}
\examples{

primer_designs <- prider("test.fasta", primer_length = 20)

}
