% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{end.position<-,PsmDetail-method}
\alias{end.position<-,PsmDetail-method}
\title{Replaces a PSM end.position}
\usage{
\S4method{end.position}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the end.position}
}
\description{
Replaces a PSM end.position
}
\author{
Jose A. Dianes
}

