% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-FileDetail.R
\docType{methods}
\name{file.size,FileDetail-method}
\alias{file.size,FileDetail-method}
\title{Returns a file size}
\usage{
\S4method{file.size}{FileDetail}(object)
}
\arguments{
\item{object}{a FileDetail}
}
\value{
the file.size
}
\description{
Returns a file size
}
\author{
Jose A. Dianes
}

