% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{file.source<-,FileDetail-method}
\alias{file.source<-,FileDetail-method}
\title{Replaces a file source}
\usage{
\S4method{file.source}{FileDetail}(object) <- value
}
\arguments{
\item{object}{a FileDetail}

\item{value}{the file.source}
}
\description{
Replaces a file source
}
\author{
Jose A. Dianes
}

