% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FileDetail}
\alias{FileDetail}
\title{Constructor for FileDetail}
\usage{
FileDetail(assay.accession, project.accession, file.name, file.type,
  file.source, file.size, download.link)
}
\arguments{
\item{assay.accession}{assay accession}

\item{project.accession}{project accession}

\item{file.name}{the name of the file}

\item{file.type}{the type of the file. e.g. RAW, PEAK, RESULT and etc}

\item{file.source}{the source of the file. e.g. SUBMITTED or GENERATED}

\item{file.size}{the size of the file}

\item{download.link}{URL for downloading the file}
}
\description{
Constructor for FileDetail
}

