% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-ContactDetail.R
\docType{methods}
\name{first.name,ContactDetail-method}
\alias{first.name,ContactDetail-method}
\title{Returns the first name of a contact}
\usage{
\S4method{first.name}{ContactDetail}(object)
}
\arguments{
\item{object}{a ContactDetail}
}
\value{
the first name
}
\description{
Returns the first name of a contact
}

