% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{first.name<-,ContactDetail-method}
\alias{first.name<-,ContactDetail-method}
\title{Replaces the first name of a contact}
\usage{
\S4method{first.name}{ContactDetail}(object) <- value
}
\arguments{
\item{object}{a ContactDetail}

\item{value}{the new first name}
}
\description{
Replaces the first name of a contact
}

