% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fromJSON.ProjectSummary}
\alias{fromJSON.ProjectSummary}
\title{Returns a ProjectSummary instance from a JSON string representation}
\usage{
fromJSON.ProjectSummary(json_str, file, method = "C")
}
\arguments{
\item{json_str}{The JSON object}

\item{json_str}{the name of a file to read the json_str from; this can also be a URL. Only one of json_str or file must be supplied.}

\item{method}{use the C implementation, or the older slower (and one day to be depricated) R implementation}

\item{unexpected.escape}{changed handling of unexpected escaped characters. Handling value should be one of "error", "skip", or "keep"; on unexpected characters issue an error, skip the character, or keep the character}
}
\value{
The ProjectSummary instance
}
\description{
Returns a ProjectSummary instance from a JSON string representation
}
\details{
TODO
}
\author{
Jose A. Dianes
}

