% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.ClusterSummary}
\alias{get.ClusterSummary}
\title{Returns a PRIDE Cluster cluster summary}
\usage{
get.ClusterSummary(id)
}
\arguments{
\item{id}{The cluster id}
}
\value{
The cluster as object
}
\description{
Returns a PRIDE Cluster cluster summary
}
\details{
TODO
}
\author{
Jose A. Dianes
}

