% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.list.assay.FileDetail}
\alias{get.list.assay.FileDetail}
\title{Returns a list of PRIDE Archive files associated with an assay}
\usage{
get.list.assay.FileDetail(assay.accession)
}
\arguments{
\item{assay.accession}{the assay accession}
}
\value{
The list of FileDetail objects
}
\description{
Returns a list of PRIDE Archive files associated with an assay
}
\details{
TODO
}
\author{
Jose A. Dianes
}

