% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assay.R
\name{get.list.AssayDetail}
\alias{get.list.AssayDetail}
\title{Returns a list of PRIDE Archive assays associated to a project}
\usage{
get.list.AssayDetail(project.accession)
}
\arguments{
\item{project.accession}{the project accession}
}
\value{
The list of AssayDetail objects
}
\description{
Returns a list of PRIDE Archive assays associated to a project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

