% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.list.ClusterSummary}
\alias{get.list.ClusterSummary}
\title{Returns a list of PRIDE Cluster cluster summaries}
\usage{
get.list.ClusterSummary(page = 1, size = 10)
}
\arguments{
\item{page}{the page number}

\item{size}{maximum number of results per page}
}
\value{
The list of ClusterSummary objects
}
\description{
Returns a list of PRIDE Cluster cluster summaries
}
\details{
TODO
}
\author{
Jose A. Dianes
}

