% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.list.project.FileDetail}
\alias{get.list.project.FileDetail}
\title{Returns a list of PRIDE Archive files associated with a project}
\usage{
get.list.project.FileDetail(project.accession)
}
\arguments{
\item{project.accession}{the project accession}
}
\value{
The list of FileDetail objects
}
\description{
Returns a list of PRIDE Archive files associated with a project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

