% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\name{get.list.ProjectSummary}
\alias{get.list.ProjectSummary}
\title{Returns a list of PRIDE Archive project summaries}
\usage{
get.list.ProjectSummary(count = 10)
}
\arguments{
\item{count}{the maximum number of projects}
}
\value{
The list of ProjectSummary objects
}
\description{
Returns a list of PRIDE Archive project summaries
}
\details{
TODO
}
\author{
Jose A. Dianes
}

