% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\name{get.list.ProteinDetail}
\alias{get.list.ProteinDetail}
\title{Returns a list of PRIDE Archive PSMs associated with a given project}
\usage{
get.list.ProteinDetail(project.accession, count = 100)
}
\arguments{
\item{count}{The maximum proteins to return from the project (deault is 1)}

\item{accession}{The project accession}
}
\value{
The list of ProteinDetail objects
}
\description{
Returns a list of PRIDE Archive PSMs associated with a given project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

