% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.ProjectSummary}
\alias{get.ProjectSummary}
\title{Returns a PRIDE Archive project}
\usage{
get.ProjectSummary(accession)
}
\arguments{
\item{accession}{The project accession}
}
\value{
The project in as object
}
\description{
Returns a PRIDE Archive project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

