% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{id,ClusterSummary-method}
\alias{id,ClusterSummary-method}
\title{Returns a cluster id}
\usage{
\S4method{id}{ClusterSummary}(object)
}
\arguments{
\item{object}{a ClusterSummary}
}
\value{
the id
}
\description{
Returns a cluster id
}
\author{
Jose A. Dianes
}

