% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{id,PsmDetail-method}
\alias{id,PsmDetail-method}
\title{Returns a PSM id}
\usage{
\S4method{id}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the ID
}
\description{
Returns a PSM id
}
\author{
Jose A. Dianes
}

