% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{id<-,ClusterSummary-method}
\alias{id<-,ClusterSummary-method}
\title{Replaces a cluster id}
\usage{
\S4method{id}{ClusterSummary}(object) <- value
}
\arguments{
\item{object}{a ClusterSummary}

\item{value}{the id}
}
\description{
Replaces a cluster id
}
\author{
Jose A. Dianes
}

