% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{id<-,PsmDetail-method}
\alias{id<-,PsmDetail-method}
\title{Replaces a PSM is}
\usage{
\S4method{id}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the id}
}
\description{
Replaces a PSM is
}
\author{
Jose A. Dianes
}

