% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{identified.spectrum.count<-,AssayDetail-method}
\alias{identified.spectrum.count<-,AssayDetail-method}
\title{Replaces an assay identified.spectrum.count}
\usage{
\S4method{identified.spectrum.count}{AssayDetail}(object) <- value
}
\arguments{
\item{object}{a AssayDetail}

\item{value}{the identified.spectrum.count}
}
\description{
Replaces an assay identified.spectrum.count
}
\author{
Jose A. Dianes
}

