% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{is_subset_of}
\alias{is_subset_of}
\title{Checks if set1 is a subset of set2. The sets are comma-delimited strings.}
\usage{
is_subset_of(set1, set2)
}
\arguments{
\item{set1}{A string}

\item{set2}{A string}
}
\value{
Boolean
}
\description{
Checks if set1 is a subset of set2. The sets are comma-delimited strings.
}
\examples{
is_subset_of("A,S,D", "A,A,S,D,B"
}
