% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{last.name<-,ContactDetail-method}
\alias{last.name<-,ContactDetail-method}
\title{Replaces the last name of a contact}
\usage{
\S4method{last.name}{ContactDetail}(object) <- value
}
\arguments{
\item{object}{a ContactDetail}

\item{value}{the new last name}
}
\description{
Replaces the last name of a contact
}

