% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.to.data.frame.ProjectSummary}
\alias{list.to.data.frame.ProjectSummary}
\title{Returns a data frame from a  list of ProjectSummary}
\usage{
list.to.data.frame.ProjectSummary(list.of.ProjectSummary)
}
\arguments{
\item{list.of.ProjectSummary}{The project summary list}
}
\value{
The project summaries as a data frame
}
\description{
Returns a data frame from a  list of ProjectSummary
}
\details{
TODO
}
\author{
Jose A. Dianes
}

