% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.to.data.frame.ProteinDetail}
\alias{list.to.data.frame.ProteinDetail}
\title{Returns a data frame from ProteinDetail list}
\usage{
list.to.data.frame.ProteinDetail(list.of.ProteinDetail)
}
\arguments{
\item{list.of.ProteinDetail}{The protein detail list}
}
\value{
The protein identification details as a data frame
}
\description{
Returns a data frame from ProteinDetail list
}
\details{
TODO
}
\author{
Jose A. Dianes
}

