% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.to.data.frame}
\alias{list.to.data.frame}
\title{Returns a data frame from a list of objects}
\usage{
list.to.data.frame(list.of.objects)
}
\arguments{
\item{list.of.objects}{The object list}
}
\value{
The objects list as a data frame
}
\description{
Returns a data frame from a list of objects
}
\details{
TODO
}
\author{
Jose A. Dianes
}

