% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{make_gephi}
\alias{make_gephi}
\title{Prepares a Gephi network file from data_list}
\usage{
make_gephi(data_list, output_gephi_file)
}
\arguments{
\item{data_list}{A list}

\item{output_gephi_file}{A string}
}
\description{
Prepares a Gephi network file from data_list
}
