% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{modifications<-,PsmDetail-method}
\alias{modifications<-,PsmDetail-method}
\title{Replaces a PSM modifications}
\usage{
\S4method{modifications}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the id}
}
\description{
Replaces a PSM modifications
}
\author{
Jose A. Dianes
}

