% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{num.assays<-,ProjectSummary-method}
\alias{num.assays<-,ProjectSummary-method}
\title{Replaces a project number of assays}
\usage{
\S4method{num.assays}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the number of assays}
}
\description{
Replaces a project number of assays
}
\author{
Jose A. Dianes
}

