% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{page.number,ProjectSummaryList-method}
\alias{page.number,ProjectSummaryList-method}
\title{Returns the page number for a given protein details list}
\usage{
\S4method{page.number}{ProjectSummaryList}(object)
}
\arguments{
\item{object}{a ProjectSummaryList instance}
}
\value{
the page number
}
\description{
Returns the page number for a given protein details list
}
\author{
Jose A. Dianes
}

