% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{page.size,ProjectSummaryList-method}
\alias{page.size,ProjectSummaryList-method}
\title{Returns the page size for a given protein details list}
\usage{
\S4method{page.size}{ProjectSummaryList}(object)
}
\arguments{
\item{object}{a ProjectSummaryList instance}
}
\value{
the page size
}
\description{
Returns the page size for a given protein details list
}
\author{
Jose A. Dianes
}

