% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{peptide.count,AssayDetail-method}
\alias{peptide.count,AssayDetail-method}
\title{Returns an assay peptide.count}
\usage{
\S4method{peptide.count}{AssayDetail}(object)
}
\arguments{
\item{object}{a AssayDetail}
}
\value{
the peptide.count
}
\description{
Returns an assay peptide.count
}
\author{
Jose A. Dianes
}

