% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assay.R
\docType{methods}
\name{peptide.count<-,AssayDetail-method}
\alias{peptide.count<-,AssayDetail-method}
\title{Replaces an assay peptide.count}
\usage{
\S4method{peptide.count}{AssayDetail}(object) <- value
}
\arguments{
\item{object}{a AssayDetail}

\item{value}{the peptide.count}
}
\description{
Replaces an assay peptide.count
}
\author{
Jose A. Dianes
}

