% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{peptide.sequence<-,Cluster-method}
\alias{peptide.sequence<-,Cluster-method}
\title{Replaces a cluster peptide.sequence}
\usage{
\S4method{peptide.sequence}{Cluster}(object) <- value
}
\arguments{
\item{object}{a Cluster}

\item{value}{the peptide.sequence}
}
\description{
Replaces a cluster peptide.sequence
}
\author{
Jose A. Dianes
}

