% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{perform_blast}
\alias{perform_blast}
\title{BLASTs the primer chunks against the original fasta. Processes the BLAST output into primer and sequence pairs.
Clusters the primers with similar targets using network analysis. Clusters the sequence targets.
Removes redundant sub-clusters. Detects overlapping clusters. Creates primer list.}
\usage{
perform_blast(data_list, min_identity = 0.8)
}
\arguments{
\item{data_list}{A list containing three tibbles: conversion table, converted sequences and primer chunks}

\item{min_identity}{A number}
}
\value{
A list containing
}
\description{
BLASTs the primer chunks against the original fasta. Processes the BLAST output into primer and sequence pairs.
Clusters the primers with similar targets using network analysis. Clusters the sequence targets.
Removes redundant sub-clusters. Detects overlapping clusters. Creates primer list.
}
