% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{plot,ClusterSearchResults-method}
\alias{plot,ClusterSearchResults-method}
\title{Plot function}
\usage{
\S4method{plot}{ClusterSearchResults}(x, y, ...)
}
\arguments{
\item{object}{a ClusterSearchResults instance}
}
\description{
Plot function
}
\author{
Jose A. Dianes
}

