% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{plot,ProteinDetailList-method}
\alias{plot,ProteinDetailList-method}
\title{Plot function}
\usage{
\S4method{plot}{ProteinDetailList}(x, y, ...)
}
\arguments{
\item{object}{a ProteinDetailList instance}
}
\description{
Plot function
}
\author{
Jose A. Dianes
}

