% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{plotresults,ClusterSearchResults-method}
\alias{plotresults,ClusterSearchResults-method}
\title{Plot function}
\usage{
\S4method{plotresults}{ClusterSearchResults}(object)
}
\arguments{
\item{object}{a ClusterSearchResults results}
}
\description{
Plot function
}
\author{
Jose A. Dianes
}

