% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{post.aa,PsmDetail-method}
\alias{post.aa,PsmDetail-method}
\title{Returns a PSM post.aa}
\usage{
\S4method{post.aa}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the post.aa
}
\description{
Returns a PSM post.aa
}
\author{
Jose A. Dianes
}

