% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{pre.aa,PsmDetail-method}
\alias{pre.aa,PsmDetail-method}
\title{Returns a PSM pre.aa}
\usage{
\S4method{pre.aa}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the pre.aa
}
\description{
Returns a PSM pre.aa
}
\author{
Jose A. Dianes
}

