% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{pre.aa<-,PsmDetail-method}
\alias{pre.aa<-,PsmDetail-method}
\title{Replaces a PSM pre.aa}
\usage{
\S4method{pre.aa}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the pre.aa}
}
\description{
Replaces a PSM pre.aa
}
\author{
Jose A. Dianes
}

