% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{prepare_blast_data}
\alias{prepare_blast_data}
\title{Prepares a fasta file for primer design by replacing simple names and chunking the sequences
using a sliding window function}
\usage{
prepare_blast_data(input_fasta, primer_length)
}
\arguments{
\item{input_fasta}{A string}

\item{primer_length}{A number}
}
\value{
A list containing three tibbles: conversion table, converted sequences and primer chunks
}
\description{
Prepares a fasta file for primer design by replacing simple names and chunking the sequences
using a sliding window function
}
\examples{
prepare_blast_data(input_fasta = "../test_align.fasta", primer_length = 20)
}
