% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{prepare_primer_table}
\alias{prepare_primer_table}
\title{Prepares a primer table for downstream analyses}
\usage{
prepare_primer_table(fasta_file, primer_length = 20)
}
\arguments{
\item{fasta_file}{A string}

\item{primer_length}{A number}
}
\value{
A list containing sequence id conversions, primer matrix
        and a list of primers with their target sequences
}
\description{
Prepares a primer table for downstream analyses
}
