% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\docType{package}
\name{prider}
\alias{prider}
\title{Prider}
\usage{
prider(input_fasta, primer_length, cluster_size)

prider(input_fasta, primer_length, cluster_size)
}
\arguments{
\item{input_fasta}{A DataFrame or a string}

\item{primer_length}{A number}
}
\value{
A list containing sequence names, sequence network, primer clusters,
        sequence clusters, cluster overlap and primer list
}
\description{
Prider implements a BLAST-based primer design algorithm for complex sequence sets.
}
\examples{

primer_designs <- prider("test.fasta", primer_length = 20)


primer_designs <- prider("test.fasta", primer_length = 20)

}
\author{
Manu Tamminen <mavatam.@utu.fi>
}
