% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{primers}
\alias{primers}
\alias{primers.default}
\alias{primers.prider}
\alias{print.primers}
\alias{[.primers}
\title{primers}
\usage{
primers(x)

\method{primers}{default}(x, ...)

\method{primers}{prider}(prider_obj)

\method{print}{primers}(primer_obj)

\method{[}{primers}(primer_obj, ix)
}
\description{
Definitions for the S3 methods for the primers classes
}
\examples{

test_fasta <- system.file("extdata", "test.fasta", package = "prider")

primer_designs <- prider(test_fasta)

primers(primer_designs)

primers(primer_designs)[1]

}
