% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{project.accession,FileDetail-method}
\alias{project.accession,FileDetail-method}
\title{Returns a file project.assay}
\usage{
\S4method{project.accession}{FileDetail}(object)
}
\arguments{
\item{object}{a FileDetail}
}
\value{
the project accession
}
\description{
Returns a file project.assay
}
\author{
Jose A. Dianes
}

