% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{project.accession,ProteinDetailList-method}
\alias{project.accession,ProteinDetailList-method}
\title{Returns a project accession for a protein list page}
\usage{
\S4method{project.accession}{ProteinDetailList}(object)
}
\arguments{
\item{object}{a ProteinDetailList instance}
}
\value{
the project accession
}
\description{
Returns a project accession for a protein list page
}
\author{
Jose A. Dianes
}

