% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{project.accession,PsmDetail-method}
\alias{project.accession,PsmDetail-method}
\title{Returns a PSM project.accession}
\usage{
\S4method{project.accession}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the project.accession
}
\description{
Returns a PSM project.accession
}
\author{
Jose A. Dianes
}

