% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assay.R
\docType{methods}
\name{project.accession<-,AssayDetail-method}
\alias{project.accession<-,AssayDetail-method}
\title{Replaces an assay project accession}
\usage{
\S4method{project.accession}{AssayDetail}(object) <- value
}
\arguments{
\item{object}{a AssayDetail}

\item{value}{the project.accession}
}
\description{
Replaces an assay project accession
}
\author{
Jose A. Dianes
}

