% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{project.accession<-,ProteinDetail-method}
\alias{project.accession<-,ProteinDetail-method}
\title{Replaces a protein project accession}
\usage{
\S4method{project.accession}{ProteinDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the project accession}
}
\description{
Replaces a protein project accession
}
\author{
Jose A. Dianes
}

