% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{project.description<-,ProjectSummary-method}
\alias{project.description<-,ProjectSummary-method}
\title{Replaces a project description}
\usage{
\S4method{project.description}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the project description}
}
\description{
Replaces a project description
}
\author{
Jose A. Dianes
}

