% Generated by roxygen2 (4.0.1): do not edit by hand
\name{project}
\alias{project}
\title{Returns a PRIDE Archive project}
\usage{
project(accession)
}
\arguments{
\item{accession}{The project accession}
}
\value{
The project in a data frame
}
\description{
Returns a PRIDE Archive project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

