% Generated by roxygen2 (4.0.1): do not edit by hand
\name{project_count}
\alias{project_count}
\title{Returns the number of public projects in PRIDE Archive}
\usage{
project_count()
}
\value{
The count of projects
}
\description{
Returns the number of public projects in PRIDE Archive
}
\details{
TODO
}
\author{
Jose A. Dianes
}

