% Generated by roxygen2 (4.0.1): do not edit by hand
\name{project_list}
\alias{project_list}
\title{Retrieves the list of projects in PRIDE Archive}
\usage{
project_list(count)
}
\arguments{
\item{count}{The maximum number of projects to retrieve in the list}
}
\value{
The list of projects in a data frame
}
\description{
Retrieves the list of projects in PRIDE Archive
}
\details{
TODO
}
\author{
Jose A. Dianes
}

