% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{class}
\name{ProjectSummary-class}
\alias{ProjectSummary-class}
\title{ProjectSummary represents a PRIDE Archive project dataset}
\description{
ProjectSummary represents a PRIDE Archive project dataset
}

