% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\name{ProjectSummary}
\alias{ProjectSummary}
\title{Constructor for ProjectSummary}
\usage{
ProjectSummary(accession, project.title, project.description, publication.date,
  num.assays, species, tissues, ptm.names, instrument.names, project.tags,
  submission.type)
}
\arguments{
\item{accession}{project accession}

\item{project.title}{the title of the project}

\item{project.description}{the description of the project}

\item{publication.date}{the date when the project was made public by PRIDE}

\item{num.assays}{the number of assays}

\item{species}{the species of the project}

\item{tissues}{the tissues of the projecte}

\item{ptm.names}{the names of the PTM for the project}

\item{instrument.names}{the names of the instruments used in the project}

\item{project.tags}{the tags for the project}

\item{submission.type}{the type of the submission, e.g. COMPLETE, PARTIAL or PRIDE}
}
\description{
Constructor for ProjectSummary
}

