% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{class}
\name{ProjectSummaryList-class}
\alias{ProjectSummaryList-class}
\title{ProjectSummaryList represents a PRIDE Archive project collection}
\description{
ProjectSummaryList represents a PRIDE Archive project collection
}

