% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{protein.accession,Cluster-method}
\alias{protein.accession,Cluster-method}
\title{Returns a cluster protein.accession}
\usage{
\S4method{protein.accession}{Cluster}(object)
}
\arguments{
\item{object}{a Cluster}
}
\value{
the protein.accession
}
\description{
Returns a cluster protein.accession
}
\author{
Jose A. Dianes
}

