% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{protein.accession,ProteinDetail-method}
\alias{protein.accession,ProteinDetail-method}
\title{Returns a protein accession}
\usage{
\S4method{protein.accession}{ProteinDetail}(object)
}
\arguments{
\item{object}{a ProjectDetail}
}
\value{
the accession
}
\description{
Returns a protein accession
}
\author{
Jose A. Dianes
}

