% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{protein.accession,PsmDetail-method}
\alias{protein.accession,PsmDetail-method}
\title{Returns a PSM protein.accession}
\usage{
\S4method{protein.accession}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the protein.accession
}
\description{
Returns a PSM protein.accession
}
\author{
Jose A. Dianes
}

