% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{protein.accession<-,Cluster-method}
\alias{protein.accession<-,Cluster-method}
\title{Replaces a cluster protein.accession}
\usage{
\S4method{protein.accession}{Cluster}(object) <- value
}
\arguments{
\item{object}{a Cluster}

\item{value}{the protein.accession}
}
\description{
Replaces a cluster protein.accession
}
\author{
Jose A. Dianes
}

