% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{protein.accession<-,PsmDetail-method}
\alias{protein.accession<-,PsmDetail-method}
\title{Replaces a PSM assay.accession}
\usage{
\S4method{protein.accession}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the protein.accession}
}
\description{
Replaces a PSM assay.accession
}
\author{
Jose A. Dianes
}

