% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{protein.count<-,AssayDetail-method}
\alias{protein.count<-,AssayDetail-method}
\title{Replaces an assay protein.count}
\usage{
\S4method{protein.count}{AssayDetail}(object) <- value
}
\arguments{
\item{object}{a AssayDetail}

\item{value}{the protein.count}
}
\description{
Replaces an assay protein.count
}
\author{
Jose A. Dianes
}

