% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{protein.list,ProteinDetailList-method}
\alias{protein.list,ProteinDetailList-method}
\title{Returns a project protein list page as a list}
\usage{
\S4method{protein.list}{ProteinDetailList}(object)
}
\arguments{
\item{object}{a ProteinDetailList instance}
}
\value{
the protein list
}
\description{
Returns a project protein list page as a list
}
\author{
Jose A. Dianes
}

