% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{protein.sequence<-,ProteinDetail-method}
\alias{protein.sequence<-,ProteinDetail-method}
\title{Replaces a protein sequence}
\usage{
\S4method{protein.sequence}{ProteinDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the sequence}
}
\description{
Replaces a protein sequence
}
\author{
Jose A. Dianes
}

