% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{protein.synonyms<-,ProteinDetail-method}
\alias{protein.synonyms<-,ProteinDetail-method}
\title{Replaces a protein synonyms}
\usage{
\S4method{protein.synonyms}{ProteinDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the synonyms}
}
\description{
Replaces a protein synonyms
}
\author{
Jose A. Dianes
}

