% Generated by roxygen2 (4.0.1): do not edit by hand
\name{protein_count}
\alias{protein_count}
\title{Returns the number of proteins for a particual public project}
\usage{
protein_count(project.accession)
}
\arguments{
\item{project.accession}{The project accession to count proteins from}
}
\value{
The count of proteins
}
\description{
Returns the number of proteins for a particual public project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

