% Generated by roxygen2 (4.0.1): do not edit by hand
\name{protein_list}
\alias{protein_list}
\title{Retrieves the list of proteins for a given PRIDE projects}
\usage{
protein_list(project.accession, count = 10)
}
\arguments{
\item{project.accession}{The project accession to list proteins from}

\item{count}{The maximum number of proteins to retrieve in the list}
}
\value{
The list of proteins in a data frame
}
\description{
Retrieves the list of proteins for a given PRIDE projects
}
\details{
TODO
}
\author{
Jose A. Dianes
}

