% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{class}
\name{ProteinDetail-class}
\alias{ProteinDetail-class}
\title{ProteinDetail represents a PRIDE Archive protein identification}
\description{
ProteinDetail represents a PRIDE Archive protein identification
}

