% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\name{ProteinDetail}
\alias{ProteinDetail}
\title{Constructor for ProteinDetail}
\usage{
ProteinDetail(protein.accession, project.accession, assay.accession,
  protein.synonyms, protein.description, protein.sequence)
}
\arguments{
\item{protein.accession}{the protein accession}

\item{project.accession}{the project accession where the protein belongs to}

\item{assay.accession}{the assay accession where the protein belongs to}

\item{protein.synonyms}{the synonyms of the protein}

\item{protein.description}{the description of the protein}

\item{protein.sequence}{the protein sequence}
}
\description{
Constructor for ProteinDetail
}

