% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ProteinDetailList-class}
\alias{ProteinDetailList-class}
\title{ProteinDetailList represents a PRIDE Archive project proteins collection}
\description{
ProteinDetailList represents a PRIDE Archive project proteins collection
}

