% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{class}
\name{PsmDetail-class}
\alias{PsmDetail-class}
\title{PsmDetail represents a PRIDE Archive peptide-spectrum matching}
\description{
PsmDetail represents a PRIDE Archive peptide-spectrum matching
}

