% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{publication.date<-,ProjectSummary-method}
\alias{publication.date<-,ProjectSummary-method}
\title{Replaces a project publication date}
\usage{
\S4method{publication.date}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the publication date}
}
\description{
Replaces a project publication date
}
\author{
Jose A. Dianes
}

