% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{query,ProjectSummaryList-method}
\alias{query,ProjectSummaryList-method}
\title{Returns the query term used to retrieve the ProjectSummaryList}
\usage{
\S4method{query}{ProjectSummaryList}(object)
}
\arguments{
\item{object}{a ProjectSummaryList instance}
}
\value{
the query term used to retrieve this ProjectSummaryList
}
\description{
Returns the query term used to retrieve the ProjectSummaryList
}
\author{
Jose A. Dianes
}

