% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{reported.id<-,PsmDetail-method}
\alias{reported.id<-,PsmDetail-method}
\title{Replaces a PSM reported.id}
\usage{
\S4method{reported.id}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the reported.id}
}
\description{
Replaces a PSM reported.id
}
\author{
Jose A. Dianes
}

