% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{results,ClusterSearchResults-method}
\alias{results,ClusterSearchResults-method}
\title{Returns a cluster search results list}
\usage{
\S4method{results}{ClusterSearchResults}(object)
}
\arguments{
\item{object}{a ClusterSearchResults results}
}
\value{
the results
}
\description{
Returns a cluster search results list
}
\author{
Jose A. Dianes
}

