% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{retention.time,PsmDetail-method}
\alias{retention.time,PsmDetail-method}
\title{Returns a PSM retention.time}
\usage{
\S4method{retention.time}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the retention.time
}
\description{
Returns a PSM retention.time
}
\author{
Jose A. Dianes
}

