% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{retention.time<-,PsmDetail-method}
\alias{retention.time<-,PsmDetail-method}
\title{Replaces a PSM retention.time}
\usage{
\S4method{retention.time}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the retention.time}
}
\description{
Replaces a PSM retention.time
}
\author{
Jose A. Dianes
}

