% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{sample_coverage}
\alias{sample_coverage}
\title{Randomly samples primer matrix to recoved the specified coverage}
\usage{
sample_coverage(primer_table, coverage = 0.9)
}
\arguments{
\item{primer_table}{A matrix}

\item{coverage}{A number}
}
\value{
A character vector containing the primer group number ids
}
\description{
Randomly samples primer matrix to recoved the specified coverage
}
