% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{search.engines,PsmDetail-method}
\alias{search.engines,PsmDetail-method}
\title{Returns a PSM search.engines}
\usage{
\S4method{search.engines}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the search.engines
}
\description{
Returns a PSM search.engines
}
\author{
Jose A. Dianes
}

