% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{search.engines.scores,PsmDetail-method}
\alias{search.engines.scores,PsmDetail-method}
\title{Returns a PSM search.engines.scores}
\usage{
\S4method{search.engines.scores}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the search.engines.scores
}
\description{
Returns a PSM search.engines.scores
}
\author{
Jose A. Dianes
}

