% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search.list.ClusterSummary}
\alias{search.list.ClusterSummary}
\title{Returns a series of PRIDE Cluster clusters
to satisify a given query. This is actually a
query filtered version of get.list.ClusterSummary}
\usage{
search.list.ClusterSummary(q, page = 1, size = 10)
}
\arguments{
\item{q}{The query terms}

\item{page}{the page number}

\item{size}{maximum number of results per page}
}
\description{
Returns a series of PRIDE Cluster clusters
to satisify a given query. This is actually a
query filtered version of get.list.ClusterSummary
}
\details{
TODO
}
\author{
Jose A. Dianes
}

