% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search.list.ProjectSummary}
\alias{search.list.ProjectSummary}
\title{Returns a series of PRIDE Archive projects
to satisify a given query. This is actually a
query filtered version of project_list}
\usage{
search.list.ProjectSummary(q, count)
}
\arguments{
\item{q}{The query terms}

\item{count}{The maximum number of search results}
}
\value{
The search results in a data frame
}
\description{
Returns a series of PRIDE Archive projects
to satisify a given query. This is actually a
query filtered version of project_list
}
\details{
TODO
}
\author{
Jose A. Dianes
}

