% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{sequence,ProteinDetail-method}
\alias{sequence,ProteinDetail-method}
\title{Returns a protein sequence}
\usage{
\S4method{sequence}{ProteinDetail}(object)
}
\arguments{
\item{object}{a ProjectDetail}
}
\value{
the sequence
}
\description{
Returns a protein sequence
}
\author{
Jose A. Dianes
}

