% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{sequence,PsmDetail-method}
\alias{sequence,PsmDetail-method}
\title{Returns a PSM sequence}
\usage{
\S4method{sequence}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the sequence
}
\description{
Returns a PSM sequence
}
\author{
Jose A. Dianes
}

