% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-ContactDetail.R
\docType{methods}
\name{show,ContactDetail-method}
\alias{show,ContactDetail-method}
\title{print out the details of the ContactDetail on screen}
\usage{
\S4method{show}{ContactDetail}(object)
}
\arguments{
\item{object}{a ContactDetail object}
}
\description{
print out the details of the ContactDetail on screen
}

