% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,ProjectSummary-method}
\alias{show,ProjectSummary-method}
\title{Show the print-out version of the content in a ProjectSummary}
\usage{
\S4method{show}{ProjectSummary}(object)
}
\arguments{
\item{object}{a given ProjectSummary}
}
\description{
Show the print-out version of the content in a ProjectSummary
}

