% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{show,ProteinDetail-method}
\alias{show,ProteinDetail-method}
\title{Show the content of a ProteinDetail}
\usage{
\S4method{show}{ProteinDetail}(object)
}
\arguments{
\item{object}{ProteinDetail object}
}
\description{
Show the content of a ProteinDetail
}

