% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\name{spectra.search.Cluster}
\alias{spectra.search.Cluster}
\title{Returns a series of PRIDE Cluster clusters
to satisify a given spectral search.}
\usage{
spectra.search.Cluster(page = 0, size = 10)
}
\arguments{
\item{page}{the page number}

\item{size}{maximum number of results per page}

\item{precursor}{The input precursor MZ}

\item{peaks}{The input spectrum peak list}
}
\description{
Returns a series of PRIDE Cluster clusters
to satisify a given spectral search.
}
\details{
TODO
}
\author{
Jose A. Dianes
}

