% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions-cluster.R
\name{spectral.search.ClusterSearchResults}
\alias{spectral.search.ClusterSearchResults}
\title{Returns a ClusterSearchResults instance to satisify a given spectral query.}
\usage{
spectral.search.ClusterSearchResults(precursor, peaks, page = 0, size = 10)
}
\arguments{
\item{precursor}{The input precursor MZ}

\item{peaks}{The input spectrum peak list}

\item{page}{the page number}

\item{size}{maximum number of results per page}
}
\description{
Returns a ClusterSearchResults instance to satisify a given spectral query.
}
\details{
TODO
}
\author{
Jose A. Dianes
}

