% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{spectrum.id,PsmDetail-method}
\alias{spectrum.id,PsmDetail-method}
\title{Returns a PSM spectrum.id}
\usage{
\S4method{spectrum.id}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the spectrum.id
}
\description{
Returns a PSM spectrum.id
}
\author{
Jose A. Dianes
}

