% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{start.postion,PsmDetail-method}
\alias{start.postion,PsmDetail-method}
\title{Returns a PSM start.postion}
\usage{
\S4method{start.postion}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the start.postion
}
\description{
Returns a PSM start.postion
}
\author{
Jose A. Dianes
}

