% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{start.postion<-,PsmDetail-method}
\alias{start.postion<-,PsmDetail-method}
\title{Replaces a PSM start.postion}
\usage{
\S4method{start.postion}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the start.postion}
}
\description{
Replaces a PSM start.postion
}
\author{
Jose A. Dianes
}

