% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{submission.type<-,ProjectSummary-method}
\alias{submission.type<-,ProjectSummary-method}
\title{Replaces the project submission type}
\usage{
\S4method{submission.type}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the submission type}
}
\description{
Replaces the project submission type
}
\author{
Jose A. Dianes
}

