% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{synonyms,ProteinDetail-method}
\alias{synonyms,ProteinDetail-method}
\title{Returns a protein synonyms}
\usage{
\S4method{synonyms}{ProteinDetail}(object)
}
\arguments{
\item{object}{a ProjectDetail}
}
\value{
the synonyms
}
\description{
Returns a protein synonyms
}
\author{
Jose A. Dianes
}

