% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{synonyms<-,ProteinDetail-method}
\alias{synonyms<-,ProteinDetail-method}
\title{Replaces a protein synonyms}
\usage{
\S4method{synonyms}{ProteinDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the synonyms}
}
\description{
Replaces a protein synonyms
}
\author{
Jose A. Dianes
}

