% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{tags,ProjectSummary-method}
\alias{tags,ProjectSummary-method}
\title{Returns a project tags}
\usage{
\S4method{tags}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project tags
}
\description{
Returns a project tags
}
\author{
Jose A. Dianes
}

