% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{tissues,ProjectSummary-method}
\alias{tissues,ProjectSummary-method}
\title{Returns a project tissues}
\usage{
\S4method{tissues}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project tissues
}
\description{
Returns a project tissues
}
\author{
Jose A. Dianes
}

