% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{tissues<-,ProjectSummary-method}
\alias{tissues<-,ProjectSummary-method}
\title{Replaces the project tissues}
\usage{
\S4method{tissues}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the tissues}
}
\description{
Replaces the project tissues
}
\author{
Jose A. Dianes
}

