% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-ContactDetail.R
\docType{methods}
\name{title,ContactDetail-method}
\alias{title,ContactDetail-method}
\title{Returns the title of a contact}
\usage{
\S4method{title}{ContactDetail}(object)
}
\arguments{
\item{object}{a ContactDetail}
}
\value{
the title
}
\description{
Returns the title of a contact
}

