% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{title,ProjectSummary-method}
\alias{title,ProjectSummary-method}
\title{Returns a project title}
\usage{
\S4method{title}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project title
}
\description{
Returns a project title
}
\author{
Jose A. Dianes
}

