% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{title<-,ContactDetail-method}
\alias{title<-,ContactDetail-method}
\title{Replaces the title of a contact}
\usage{
\S4method{title}{ContactDetail}(object) <- value
}
\arguments{
\item{object}{a ContactDetail}

\item{value}{the new title}
}
\description{
Replaces the title of a contact
}

