% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{title<-,ProjectSummary-method}
\alias{title<-,ProjectSummary-method}
\title{Replaces a project title}
\usage{
\S4method{title}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the title}
}
\description{
Replaces a project title
}
\author{
Jose A. Dianes
}

