% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assay.R
\docType{methods}
\name{total.spectrum.count,AssayDetail-method}
\alias{total.spectrum.count,AssayDetail-method}
\title{Returns an assay total.spectrum.count}
\usage{
\S4method{total.spectrum.count}{AssayDetail}(object)
}
\arguments{
\item{object}{a AssayDetail}
}
\value{
the total.spectrum.count
}
\description{
Returns an assay total.spectrum.count
}
\author{
Jose A. Dianes
}

