% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{unique.peptide.count,AssayDetail-method}
\alias{unique.peptide.count,AssayDetail-method}
\title{Returns an assay unique.peptide.count}
\usage{
\S4method{unique.peptide.count}{AssayDetail}(object)
}
\arguments{
\item{object}{a AssayDetail}
}
\value{
the unique.peptide.count
}
\description{
Returns an assay unique.peptide.count
}
\author{
Jose A. Dianes
}

