% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{unique.peptide.count<-,AssayDetail-method}
\alias{unique.peptide.count<-,AssayDetail-method}
\title{Replaces an assay unique.peptide.count}
\usage{
\S4method{unique.peptide.count}{AssayDetail}(object) <- value
}
\arguments{
\item{object}{a AssayDetail}

\item{value}{the unique.peptide.count}
}
\description{
Replaces an assay unique.peptide.count
}
\author{
Jose A. Dianes
}

