% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{accession,ProjectSummary-method}
\alias{accession,ProjectSummary-method}
\title{Returns a project accession}
\usage{
\S4method{accession}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the accession
}
\description{
Returns a project accession
}
\author{
Jose A. Dianes
}

