% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.data.frame,ProteinDetail-method}
\alias{as.data.frame,ProteinDetail-method}
\title{Returns a data frame from ProteinDetail inputs}
\usage{
\S4method{as.data.frame}{ProteinDetail}(object, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{row.names}{optional row names}

\item{x}{The protein detail inputs}
}
\value{
The protein identification details as a data frame
}
\description{
Returns a data frame from ProteinDetail inputs
}
\details{
TODO
}
\author{
Jose A. Dianes
}

