% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assay.R
\name{as.data.frame.AssayDetail}
\alias{as.data.frame.AssayDetail}
\title{Returns a data frame from AssayDetail inputs}
\usage{
\method{as.data.frame}{AssayDetail}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{The assay details}

\item{row.names}{optional row names}
}
\value{
The assay details as a data frame
}
\description{
Returns a data frame from AssayDetail inputs
}
\details{
TODO
}
\author{
Jose A. Dianes
}

