% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.Cluster}
\alias{as.data.frame.Cluster}
\title{Returns a data frame from Cluster inputs}
\usage{
\method{as.data.frame}{Cluster}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{The cluster summaries}

\item{row.names}{optional row names}

\item{optional}{optional}
}
\value{
The cluster summaries as a data frame
}
\description{
Returns a data frame from Cluster inputs
}
\details{
TODO
}
\author{
Jose A. Dianes
}

