% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.FileDetail}
\alias{as.data.frame.FileDetail}
\title{Returns a data frame from FileDetail inputs}
\usage{
as.data.frame.FileDetail(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{The file details}

\item{row.names}{optional row names}
}
\value{
The file details as a data frame
}
\description{
Returns a data frame from FileDetail inputs
}
\details{
TODO
}
\author{
Jose A. Dianes
}

