% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.data.frame.ProjectSummary}
\alias{as.data.frame.ProjectSummary}
\title{Returns a ProjectSummary instance from a JSON string representation}
\usage{
\method{as.data.frame}{ProjectSummary}(x, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{json_str}{The JSON object}

\item{x}{The project summaries}

\item{row.names}{optional row names}
}
\value{
The project summaries as a data frame
}
\description{
Returns a ProjectSummary instance from a JSON string representation
}
\details{
TODO
}
\author{
Jose A. Dianes
}

