% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.ProjectSummary}
\alias{as.data.frame.ProjectSummary}
\title{Returns a data frame from ProjectSummary inputs}
\usage{
\method{as.data.frame}{ProjectSummary}(x, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{x}{The project summaries}

\item{row.names}{optional row names}
}
\value{
The project summaries as a data frame
}
\description{
Returns a data frame from ProjectSummary inputs
}
\details{
TODO
}
\author{
Jose A. Dianes
}

