% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{assay.accession<-,PsmDetail-method}
\alias{assay.accession<-,PsmDetail-method}
\title{Replaces a PSM assay.accession}
\usage{
\S4method{assay.accession}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the assay.accession}
}
\description{
Replaces a PSM assay.accession
}
\author{
Jose A. Dianes
}

