% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{average.precursor.mz<-,Cluster-method}
\alias{average.precursor.mz<-,Cluster-method}
\title{Replaces a cluster average.precursor.mz}
\usage{
\S4method{average.precursor.mz}{Cluster}(object) <- value
}
\arguments{
\item{object}{a Cluster}

\item{value}{the average.precursor.mz}
}
\description{
Replaces a cluster average.precursor.mz
}
\author{
Jose A. Dianes
}

