% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{calculated.mz<-,PsmDetail-method}
\alias{calculated.mz<-,PsmDetail-method}
\title{Replaces a PSM calculated.mz}
\usage{
\S4method{calculated.mz}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the calculated.mz}
}
\description{
Replaces a PSM calculated.mz
}
\author{
Jose A. Dianes
}

