% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chunker}
\alias{chunker}
\title{Creates all primer candidates for a group of sequences using a sliding window.}
\usage{
chunker(seq_table, window_size = 20L)
}
\arguments{
\item{seq_table}{A DataFrame}

\item{window_size}{An integer}
}
\value{
A DataFrame containing columns for sequence ids, 
        indexes joined ids and indexes and the primer sequences
}
\description{
Creates all primer candidates for a group of sequences using a sliding window.
}
