% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chunker}
\alias{chunker}
\title{Create chunks from a DNA sequence using a sliding window}
\usage{
chunker(seq_table, window_size = 20L)
}
\arguments{
\item{seq_table}{A DataFrame}

\item{window_size}{An integer}
}
\description{
Create chunks from a DNA sequence using a sliding window
}
