% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{cluster.quality,Cluster-method}
\alias{cluster.quality,Cluster-method}
\title{Returns a cluster cluster.quality}
\usage{
\S4method{cluster.quality}{Cluster}(object)
}
\arguments{
\item{object}{a Cluster}
}
\value{
the cluster.quality
}
\description{
Returns a cluster cluster.quality
}
\author{
Jose A. Dianes
}

