% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{cluster.quality<-,Cluster-method}
\alias{cluster.quality<-,Cluster-method}
\title{Replaces a cluster cluster.quality}
\usage{
\S4method{cluster.quality}{Cluster}(object) <- value
}
\arguments{
\item{object}{a Cluster}

\item{value}{the cluster.quality}
}
\description{
Replaces a cluster cluster.quality
}
\author{
Jose A. Dianes
}

