% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ClusterSearchResults-class}
\alias{ClusterSearchResults-class}
\title{ClusterSearchResults represents a PRIDE Cluster project summary result set}
\description{
ClusterSearchResults represents a PRIDE Cluster project summary result set
}

