% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count.AssayDetail}
\alias{count.AssayDetail}
\title{Returns the number of assays associated with a project}
\usage{
count.AssayDetail(project.accession)
}
\value{
The count of assays
}
\description{
Returns the number of assays associated with a project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

