% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count.project.FileDetail}
\alias{count.project.FileDetail}
\title{Returns the number of files associated with a project}
\usage{
count.project.FileDetail(project.accession)
}
\arguments{
\item{project.accession}{the project accession}
}
\value{
The count of files
}
\description{
Returns the number of files associated with a project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

