% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\name{count.PsmDetail}
\alias{count.PsmDetail}
\title{Returns the number of PSM for a particual public project}
\usage{
count.PsmDetail(project.accession)
}
\arguments{
\item{project.accession}{The project accession to count PSM from}
}
\value{
The count of PSM
}
\description{
Returns the number of PSM for a particual public project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

