% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{download.link<-,FileDetail-method}
\alias{download.link<-,FileDetail-method}
\title{Replaces a file download.link}
\usage{
\S4method{download.link}{FileDetail}(object) <- value
}
\arguments{
\item{object}{a FileDetail}

\item{value}{the download.link}
}
\description{
Replaces a file download.link
}
\author{
Jose A. Dianes
}

