% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-ContactDetail.R
\docType{methods}
\name{email,ContactDetail-method}
\alias{email,ContactDetail-method}
\title{Returns the email of a contact}
\usage{
\S4method{email}{ContactDetail}(object)
}
\arguments{
\item{object}{a ContactDetail}
}
\value{
the email
}
\description{
Returns the email of a contact
}

