% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{email<-,ContactDetail-method}
\alias{email<-,ContactDetail-method}
\title{Replaces the email of a contact}
\usage{
\S4method{email}{ContactDetail}(object) <- value
}
\arguments{
\item{object}{a ContactDetail}

\item{value}{the new email}
}
\description{
Replaces the email of a contact
}

