% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{experimental.mz<-,PsmDetail-method}
\alias{experimental.mz<-,PsmDetail-method}
\title{Replaces a PSM experimental.mz}
\usage{
\S4method{experimental.mz}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the experimental.mz}
}
\description{
Replaces a PSM experimental.mz
}
\author{
Jose A. Dianes
}

