% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-FileDetail.R
\docType{methods}
\name{file.name,FileDetail-method}
\alias{file.name,FileDetail-method}
\title{Returns a file name}
\usage{
\S4method{file.name}{FileDetail}(object)
}
\arguments{
\item{object}{a FileDetail}
}
\value{
the file.name
}
\description{
Returns a file name
}
\author{
Jose A. Dianes
}

