% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{file.source,FileDetail-method}
\alias{file.source,FileDetail-method}
\title{Returns a file source}
\usage{
\S4method{file.source}{FileDetail}(object)
}
\arguments{
\item{object}{a FileDetail}
}
\value{
the file.source
}
\description{
Returns a file source
}
\author{
Jose A. Dianes
}

