% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{file.type,FileDetail-method}
\alias{file.type,FileDetail-method}
\title{Returns a file type}
\usage{
\S4method{file.type}{FileDetail}(object)
}
\arguments{
\item{object}{a FileDetail}
}
\value{
the file.type
}
\description{
Returns a file type
}
\author{
Jose A. Dianes
}

