% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-FileDetail.R
\docType{methods}
\name{file.type<-,FileDetail-method}
\alias{file.type<-,FileDetail-method}
\title{Replaces a file type}
\usage{
\S4method{file.type}{FileDetail}(object) <- value
}
\arguments{
\item{object}{a FileDetail}

\item{value}{the file.type}
}
\description{
Replaces a file type
}
\author{
Jose A. Dianes
}

