% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.AssayDetail}
\alias{get.AssayDetail}
\title{Returns a PRIDE Archive assay}
\usage{
get.AssayDetail(assay.accession)
}
\arguments{
\item{accession}{The assay accession}
}
\value{
The assay as object of AssayDetail
}
\description{
Returns a PRIDE Archive assay
}
\details{
TODO
}
\author{
Jose A. Dianes
}

