% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions-cluster.R
\name{get.Cluster}
\alias{get.Cluster}
\title{Returns a PRIDE Cluster cluster}
\usage{
get.Cluster(id)
}
\arguments{
\item{id}{The cluster id}
}
\value{
The cluster as object
}
\description{
Returns a PRIDE Cluster cluster
}
\details{
TODO
}
\author{
Jose A. Dianes
}

