% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\name{get.list.Cluster}
\alias{get.list.Cluster}
\title{Returns a list of PRIDE Cluster cluster summaries}
\usage{
get.list.Cluster(page = 0, size = 10)
}
\arguments{
\item{page}{the page number}

\item{size}{maximum number of results per page}
}
\value{
The list of Cluster objects
}
\description{
Returns a list of PRIDE Cluster cluster summaries
}
\details{
TODO
}
\author{
Jose A. Dianes
}

