% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.list.ProjectSummary}
\alias{get.list.ProjectSummary}
\title{Returns a list of PRIDE Archive project summaries}
\usage{
get.list.ProjectSummary(count = 1)
}
\arguments{
\item{count}{the maximum number of projects}
}
\value{
The list of ProjectSummary objects
}
\description{
Returns a list of PRIDE Archive project summaries
}
\details{
TODO
}
\author{
Jose A. Dianes
}

