% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.list.ProteinDetail}
\alias{get.list.ProteinDetail}
\title{Returns a list of PRIDE Archive PSMs associated with a given project}
\usage{
get.list.ProteinDetail(project.accession, count = 1)
}
\arguments{
\item{accession}{The project accession}
}
\value{
The list of ProteinDetail objects
}
\description{
Returns a list of PRIDE Archive PSMs associated with a given project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

