% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.list.PsmDetail}
\alias{get.list.PsmDetail}
\title{Returns a list of PRIDE Archive PSMs associated with a given project}
\usage{
get.list.PsmDetail(project.accession, count = 100)
}
\arguments{
\item{count}{The maximum PSMs to return from the project (deault is 1)}

\item{accession}{The project accession}
}
\value{
The list of PSM objects
}
\description{
Returns a list of PRIDE Archive PSMs associated with a given project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

