% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{id<-,Cluster-method}
\alias{id<-,Cluster-method}
\title{Replaces a cluster id}
\usage{
\S4method{id}{Cluster}(object) <- value
}
\arguments{
\item{object}{a Cluster}

\item{value}{the id}
}
\description{
Replaces a cluster id
}
\author{
Jose A. Dianes
}

