% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{id<-,PsmDetail-method}
\alias{id<-,PsmDetail-method}
\title{Replaces a PSM is}
\usage{
\S4method{id}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the id}
}
\description{
Replaces a PSM is
}
\author{
Jose A. Dianes
}

