% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{instrument.names,ProjectSummary-method}
\alias{instrument.names,ProjectSummary-method}
\title{Returns a project instrument names}
\usage{
\S4method{instrument.names}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project instrument names
}
\description{
Returns a project instrument names
}
\author{
Jose A. Dianes
}

