% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{instrument.names<-,ProjectSummary-method}
\alias{instrument.names<-,ProjectSummary-method}
\title{Replaces the project instrument nanmes}
\usage{
\S4method{instrument.names}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the instrument names}
}
\description{
Replaces the project instrument nanmes
}
\author{
Jose A. Dianes
}

