% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-ContactDetail.R
\docType{methods}
\name{last.name,ContactDetail-method}
\alias{last.name,ContactDetail-method}
\title{Returns the last name of a contact}
\usage{
\S4method{last.name}{ContactDetail}(object)
}
\arguments{
\item{object}{a ContactDetail}
}
\value{
the last name
}
\description{
Returns the last name of a contact
}

