% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.ProteinDetailList}
\alias{list.ProteinDetailList}
\title{Returns a list of PRIDE Archive PSMs associated with a given project}
\usage{
list.ProteinDetailList(project.accession, page.number = 0, page.size = 100)
}
\arguments{
\item{accession}{The project accession}

\item{count}{The maximum proteins to return from the project (deault is 1)}
}
\value{
The list of ProteinDetail objects
}
\description{
Returns a list of PRIDE Archive PSMs associated with a given project
}
\details{
TODO
}
\author{
Jose A. Dianes
}

