% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{max.ratio,Cluster-method}
\alias{max.ratio,Cluster-method}
\title{Returns a cluster max.ratio}
\usage{
\S4method{max.ratio}{Cluster}(object)
}
\arguments{
\item{object}{a Cluster}
}
\value{
the max.ratio
}
\description{
Returns a cluster max.ratio
}
\author{
Jose A. Dianes
}

