% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{max.ratio<-,Cluster-method}
\alias{max.ratio<-,Cluster-method}
\title{Replaces a cluster max.ratio}
\usage{
\S4method{max.ratio}{Cluster}(object) <- value
}
\arguments{
\item{object}{a Cluster}

\item{value}{the max.ratio}
}
\description{
Replaces a cluster max.ratio
}
\author{
Jose A. Dianes
}

