% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{modifications,PsmDetail-method}
\alias{modifications,PsmDetail-method}
\title{Returns a PSM modifications}
\usage{
\S4method{modifications}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the modifications
}
\description{
Returns a PSM modifications
}
\author{
Jose A. Dianes
}

