% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{num.assays,ProjectSummary-method}
\alias{num.assays,ProjectSummary-method}
\title{Returns a project number of assays}
\usage{
\S4method{num.assays}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the number of assays
}
\description{
Returns a project number of assays
}
\author{
Jose A. Dianes
}

