% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{num.spectra,Cluster-method}
\alias{num.spectra,Cluster-method}
\title{Returns a cluster num.spectra}
\usage{
\S4method{num.spectra}{Cluster}(object)
}
\arguments{
\item{object}{a Cluster}
}
\value{
the num.spectra
}
\description{
Returns a cluster num.spectra
}
\author{
Jose A. Dianes
}

