% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{num.spectra<-,Cluster-method}
\alias{num.spectra<-,Cluster-method}
\title{Replaces a cluster num.spectra}
\usage{
\S4method{num.spectra}{Cluster}(object) <- value
}
\arguments{
\item{object}{a Cluster}

\item{value}{the num.spectra}
}
\description{
Replaces a cluster num.spectra
}
\author{
Jose A. Dianes
}

