% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{page.number,ClusterSearchResults-method}
\alias{page.number,ClusterSearchResults-method}
\title{Returns the page number for a given search results}
\usage{
\S4method{page.number}{ClusterSearchResults}(object)
}
\arguments{
\item{object}{a ClusterSearchResults instance}
}
\value{
the page number
}
\description{
Returns the page number for a given search results
}
\author{
Jose A. Dianes
}

