% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{page.number,ProteinDetailList-method}
\alias{page.number,ProteinDetailList-method}
\title{Returns the page number for a given protein details list}
\usage{
\S4method{page.number}{ProteinDetailList}(object)
}
\arguments{
\item{object}{a ProteinDetailList instance}
}
\value{
the page number
}
\description{
Returns the page number for a given protein details list
}
\author{
Jose A. Dianes
}

