% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{page.size,ClusterSearchResults-method}
\alias{page.size,ClusterSearchResults-method}
\title{Returns the page size for a given search results}
\usage{
\S4method{page.size}{ClusterSearchResults}(object)
}
\arguments{
\item{object}{a ClusterSearchResults page size}
}
\value{
the page size
}
\description{
Returns the page size for a given search results
}
\author{
Jose A. Dianes
}

