% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{page.size,ProteinDetailList-method}
\alias{page.size,ProteinDetailList-method}
\title{Returns the page size for a given protein details list}
\usage{
\S4method{page.size}{ProteinDetailList}(object)
}
\arguments{
\item{object}{a ProteinDetailList instance}
}
\value{
the page size
}
\description{
Returns the page size for a given protein details list
}
\author{
Jose A. Dianes
}

