% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{peptide.sequence,Cluster-method}
\alias{peptide.sequence,Cluster-method}
\title{Returns a cluster peptide.sequence}
\usage{
\S4method{peptide.sequence}{Cluster}(object)
}
\arguments{
\item{object}{a Cluster}
}
\value{
the peptide.sequence
}
\description{
Returns a cluster peptide.sequence
}
\author{
Jose A. Dianes
}

