% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot,ClusterSearchResults-method}
\alias{plot,ClusterSearchResults-method}
\title{Plot function}
\usage{
\S4method{plot}{ClusterSearchResults}(x, y, ...)
}
\arguments{
\item{object}{a ClusterSearchResults results}
}
\description{
Plot function
}
\author{
Jose A. Dianes
}

