% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{post.aa<-,PsmDetail-method}
\alias{post.aa<-,PsmDetail-method}
\title{Replaces a PSM post.aa}
\usage{
\S4method{post.aa}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the post.aa}
}
\description{
Replaces a PSM post.aa
}
\author{
Jose A. Dianes
}

