% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{prepare_primer_df}
\alias{prepare_primer_df}
\title{Prepare a primer table for downstream analyses}
\usage{
prepare_primer_df(
  input_fasta,
  primer_length = 20,
  GCcheck = FALSE,
  GCsimilarity = 0.1
)
}
\arguments{
\item{input_fasta}{A string or a DataFrame containing Id and Seq columns.}

\item{primer_length}{A number specifying length for the designed primers.}

\item{GCcheck}{A logical; check the GC contents of the primer halves.}

\item{GCsimilarity}{A number; if GCcheck is performed, this parameter
determines the maximum proportional GC content
difference between the primer halves.}
}
\value{
A list containing sequence id conversions, primer matrix
        and a list of primers with their target sequences.
}
\description{
Prepare a primer table for downstream analyses
}
