% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{prepare_primer_table}
\alias{prepare_primer_table}
\title{Prepare a primer table for downstream analyses}
\usage{
prepare_primer_table(input_fasta, primer_length = 20)
}
\arguments{
\item{input_fasta}{A string or a DataFrame containing Id and Seq columns}

\item{primer_length}{A number}
}
\value{
A list containing sequence id conversions, primer matrix
        and a list of primers with their target sequences
}
\description{
Prepare a primer table for downstream analyses
}
