% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\docType{package}
\name{prider}
\alias{prider}
\title{Prider}
\usage{
prider(input_fasta, primer_length, min_identity = 0.95)
}
\arguments{
\item{input_fasta}{A string}

\item{primer_length}{A number}
}
\value{
A list containing three tibbles: conversion table, converted sequences and primer chunks
}
\description{
Prider implements a BLAST-based primer design algorithm for complex sequence sets.
}
\author{
Manu Tamminen <mavatam.@utu.fi>
}
