% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\docType{package}
\name{prider}
\alias{prider}
\alias{print.prider}
\alias{plot.prider}
\title{Prider}
\usage{
prider(
  fasta_file,
  primer_length = 20,
  minimum_primer_group_size = 10,
  draws = 10
)

\method{print}{prider}(prider_obj)

\method{plot}{prider}(prider_obj)
}
\arguments{
\item{fasta_file}{A string}

\item{primer_length}{A number}

\item{minimum_primer_group_size}{A number}

\item{draws}{A number}
}
\value{
A list containing a sequence conversion table and 
        a primer coverage table
}
\description{
Prider implements a BLAST-based primer design algorithm for complex sequence sets.

Prepare a (nearly) optimal primer coverage of the sample set
}
\examples{

test_fasta <- system.file("extdata", "test.fasta", package = "prider")

primer_designs <- prider(test_fasta)

}
\author{
Manu Tamminen <mavatam.@utu.fi>
}
