% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{primers}
\alias{primers}
\alias{primers.default}
\alias{primers.prider}
\alias{print.primers}
\alias{[.primers}
\title{primers}
\usage{
primers(x)

\method{primers}{default}(x, ...)

\method{primers}{prider}(prider_obj)

\method{print}{primers}(primer_obj)

\method{[}{primers}(primer_obj, ix)
}
\arguments{
\item{prider_obj}{A list}

\item{primer_obj}{A list}
}
\description{
primers
}
