% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{primers}
\alias{primers}
\alias{primers.default}
\alias{primers.prider}
\alias{print.primers}
\alias{[.primers}
\title{primers}
\usage{
primers(prider_obj)

\method{primers}{default}(prider_obj)

\method{primers}{prider}(prider_obj)

\method{print}{primers}(x, ...)

\method{[}{primers}(primer_obj, ix)
}
\arguments{
\item{prider_obj}{An object from prider function.}

\item{x}{An object from sequence function.}

\item{...}{Other arguments.}

\item{primer_obj}{An object from sequence function.}

\item{ix}{A number. The number of the primer cluster.}
}
\value{
primer_obj
}
\description{
Definitions for the S3 methods for the primers classes
}
\examples{

test_fasta <- system.file("extdata", "test.fasta", package = "prider")

primer_designs <- prider(test_fasta)

primers(primer_designs)

primers(primer_designs)[1]

}
