% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{project.accession,AssayDetail-method}
\alias{project.accession,AssayDetail-method}
\title{Returns an assay project.assay}
\usage{
\S4method{project.accession}{AssayDetail}(object)
}
\arguments{
\item{object}{a AssayDetail}
}
\value{
the project accession
}
\description{
Returns an assay project.assay
}
\author{
Jose A. Dianes
}

