% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{project.accession,ProteinDetail-method}
\alias{project.accession,ProteinDetail-method}
\title{Returns a protein project accession}
\usage{
\S4method{project.accession}{ProteinDetail}(object)
}
\arguments{
\item{object}{a ProjectDetail}
}
\value{
the project accession
}
\description{
Returns a protein project accession
}
\author{
Jose A. Dianes
}

