% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-FileDetail.R
\docType{methods}
\name{project.accession<-,FileDetail-method}
\alias{project.accession<-,FileDetail-method}
\title{Replaces a file project accession}
\usage{
\S4method{project.accession}{FileDetail}(object) <- value
}
\arguments{
\item{object}{a FileDetail}

\item{value}{the project.accession}
}
\description{
Replaces a file project accession
}
\author{
Jose A. Dianes
}

