% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{project.accession<-,PsmDetail-method}
\alias{project.accession<-,PsmDetail-method}
\title{Replaces a PSM project.accession}
\usage{
\S4method{project.accession}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the project.accession}
}
\description{
Replaces a PSM project.accession
}
\author{
Jose A. Dianes
}

