% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{protein.accession<-,ProteinDetail-method}
\alias{protein.accession<-,ProteinDetail-method}
\title{Replaces a protein accession}
\usage{
\S4method{protein.accession}{ProteinDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the accession}
}
\description{
Replaces a protein accession
}
\author{
Jose A. Dianes
}

