% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{protein.description,ProteinDetail-method}
\alias{protein.description,ProteinDetail-method}
\title{Returns a protein description}
\usage{
\S4method{protein.description}{ProteinDetail}(object)
}
\arguments{
\item{object}{a ProjectDetail}
}
\value{
the description
}
\description{
Returns a protein description
}
\author{
Jose A. Dianes
}

