% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{methods}
\name{protein.description<-,ProteinDetail-method}
\alias{protein.description<-,ProteinDetail-method}
\title{Replaces a protein description}
\usage{
\S4method{protein.description}{ProteinDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the description}
}
\description{
Replaces a protein description
}
\author{
Jose A. Dianes
}

