% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{protein.sequence,ProteinDetail-method}
\alias{protein.sequence,ProteinDetail-method}
\title{Returns a protein sequence}
\usage{
\S4method{protein.sequence}{ProteinDetail}(object)
}
\arguments{
\item{object}{a ProjectDetail}
}
\value{
the sequence
}
\description{
Returns a protein sequence
}
\author{
Jose A. Dianes
}

