% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protein.R
\docType{class}
\name{ProteinDetailList-class}
\alias{ProteinDetailList-class}
\title{ProteinDetailList represents a PRIDE Archive project proteins collection}
\description{
ProteinDetailList represents a PRIDE Archive project proteins collection
}

