% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{ptm.names,ProjectSummary-method}
\alias{ptm.names,ProjectSummary-method}
\title{Returns a project modification names}
\usage{
\S4method{ptm.names}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project modification names
}
\description{
Returns a project modification names
}
\author{
Jose A. Dianes
}

