% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ptm.names<-,ProjectSummary-method}
\alias{ptm.names<-,ProjectSummary-method}
\title{Replaces the project PTMs}
\usage{
\S4method{ptm.names}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the PTMs}
}
\description{
Replaces the project PTMs
}
\author{
Jose A. Dianes
}

