% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{publication.date,ProjectSummary-method}
\alias{publication.date,ProjectSummary-method}
\title{Returns a project publication date}
\usage{
\S4method{publication.date}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project publication date
}
\description{
Returns a project publication date
}
\author{
Jose A. Dianes
}

