% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{reported.id,PsmDetail-method}
\alias{reported.id,PsmDetail-method}
\title{Returns a PSM reported.id}
\usage{
\S4method{reported.id}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the reported.id
}
\description{
Returns a PSM reported.id
}
\author{
Jose A. Dianes
}

