% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psm.R
\docType{methods}
\name{reported.id<-,PsmDetail-method}
\alias{reported.id<-,PsmDetail-method}
\title{Replaces a PSM reported.id}
\usage{
\S4method{reported.id}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the reported.id}
}
\description{
Replaces a PSM reported.id
}
\author{
Jose A. Dianes
}

