% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions-cluster.R
\name{search.ClusterSearchResults}
\alias{search.ClusterSearchResults}
\title{Returns a ClusterSearchResults instance to satisify a given peptide query.}
\usage{
search.ClusterSearchResults(q = "", page = 0, size = 10)
}
\arguments{
\item{q}{The query terms}

\item{page}{the page number}

\item{size}{maximum number of results per page}
}
\description{
Returns a ClusterSearchResults instance to satisify a given peptide query.
}
\details{
TODO
}
\author{
Jose A. Dianes
}

