% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{search.engines<-,PsmDetail-method}
\alias{search.engines<-,PsmDetail-method}
\title{Replaces a PSM search.engines}
\usage{
\S4method{search.engines}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the search.engines}
}
\description{
Replaces a PSM search.engines
}
\author{
Jose A. Dianes
}

