% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\name{search.list.ProjectSummary}
\alias{search.list.ProjectSummary}
\title{Returns a series of PRIDE Archive projects
to satisify a given query. This is actually a
query filtered version of project_list}
\usage{
search.list.ProjectSummary(q, page.number = 0, page.size = 100)
}
\arguments{
\item{q}{The query terms}

\item{count}{The maximum number of search results}
}
\value{
The search results in a list of objects
}
\description{
Returns a series of PRIDE Archive projects
to satisify a given query. This is actually a
query filtered version of project_list
}
\details{
TODO
}
\author{
Jose A. Dianes
}

