% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{sequence<-,ProteinDetail-method}
\alias{sequence<-,ProteinDetail-method}
\title{Replaces a protein sequence}
\usage{
\S4method{sequence}{ProteinDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the sequence}
}
\description{
Replaces a protein sequence
}
\author{
Jose A. Dianes
}

