% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{sequence<-,PsmDetail-method}
\alias{sequence<-,PsmDetail-method}
\title{Replaces a PSM sequence}
\usage{
\S4method{sequence}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the sequence}
}
\description{
Replaces a PSM sequence
}
\author{
Jose A. Dianes
}

