% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{sequences}
\alias{sequences}
\alias{sequences.default}
\alias{sequences.prider}
\alias{print.sequences}
\alias{[.sequences}
\title{sequences}
\usage{
sequences(x)

\method{sequences}{default}(x, ...)

\method{sequences}{prider}(prider_obj)

\method{print}{sequences}(sequence_obj)

\method{[}{sequences}(sequence_obj, ix)
}
\description{
Definitions for the S3 methods for the sequences classes
}
\examples{

test_fasta <- system.file("extdata", "test.fasta", package = "prider")

primer_designs <- prider(test_fasta)

sequences(primer_designs)

sequences(primer_designs)[1]

}
