% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prider.R
\name{sequences}
\alias{sequences}
\alias{sequences.default}
\alias{sequences.prider}
\alias{print.sequences}
\alias{[.sequences}
\title{sequences}
\usage{
sequences(prider_obj)

\method{sequences}{default}(prider_obj)

\method{sequences}{prider}(prider_obj)

\method{print}{sequences}(x, ...)

\method{[}{sequences}(sequence_obj, ix)
}
\arguments{
\item{prider_obj}{An object from prider function.}

\item{x}{An object from sequence function.}

\item{...}{Other arguments.}

\item{sequence_obj}{An object from sequence function.}

\item{ix}{A number. The number of the primer cluster.}
}
\value{
sequence_obj
}
\description{
Definitions for the S3 methods for the sequences classes
}
\examples{

test_fasta <- system.file("extdata", "test.fasta", package = "prider")

primer_designs <- prider(test_fasta)

sequences(primer_designs)

sequences(primer_designs)[1]

}
