% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,ProteinDetail-method}
\alias{show,ProteinDetail-method}
\title{Show the content of a ProteinDetail}
\usage{
\S4method{show}{ProteinDetail}(object)
}
\arguments{
\item{object}{ProteinDetail object}
}
\description{
Show the content of a ProteinDetail
}

