% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{species,ProjectSummary-method}
\alias{species,ProjectSummary-method}
\title{Returns a project species}
\usage{
\S4method{species}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project species
}
\description{
Returns a project species
}
\author{
Jose A. Dianes
}

