% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{species<-,ProjectSummary-method}
\alias{species<-,ProjectSummary-method}
\title{Replaces the project species}
\usage{
\S4method{species}{ProjectSummary}(object) <- value
}
\arguments{
\item{object}{a ProjectSummary}

\item{value}{the species}
}
\description{
Replaces the project species
}
\author{
Jose A. Dianes
}

