% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spectrum.id<-,PsmDetail-method}
\alias{spectrum.id<-,PsmDetail-method}
\title{Replaces a PSM spectrum.id}
\usage{
\S4method{spectrum.id}{PsmDetail}(object) <- value
}
\arguments{
\item{object}{a ProteinDetail}

\item{value}{the spectrum.id}
}
\description{
Replaces a PSM spectrum.id
}
\author{
Jose A. Dianes
}

