% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{start.postion,PsmDetail-method}
\alias{start.postion,PsmDetail-method}
\title{Returns a PSM start.postion}
\usage{
\S4method{start.postion}{PsmDetail}(object)
}
\arguments{
\item{object}{a PsmDetail}
}
\value{
the start.postion
}
\description{
Returns a PSM start.postion
}
\author{
Jose A. Dianes
}

