% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{submission.type,ProjectSummary-method}
\alias{submission.type,ProjectSummary-method}
\title{Returns a project submission type}
\usage{
\S4method{submission.type}{ProjectSummary}(object)
}
\arguments{
\item{object}{a ProjectSummary}
}
\value{
the project submission type
}
\description{
Returns a project submission type
}
\author{
Jose A. Dianes
}

