% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{synonyms,ProteinDetail-method}
\alias{synonyms,ProteinDetail-method}
\title{Returns a protein synonyms}
\usage{
\S4method{synonyms}{ProteinDetail}(object)
}
\arguments{
\item{object}{a ProjectDetail}
}
\value{
the synonyms
}
\description{
Returns a protein synonyms
}
\author{
Jose A. Dianes
}

