% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{total.results,ClusterSearchResults-method}
\alias{total.results,ClusterSearchResults-method}
\title{Returns the total number of clusters for a given search results}
\usage{
\S4method{total.results}{ClusterSearchResults}(object)
}
\arguments{
\item{object}{a ClusterSearchResults total results}
}
\value{
the total number of results
}
\description{
Returns the total number of clusters for a given search results
}
\author{
Jose A. Dianes
}

